tech: improving tasks handling, tests and logging
